+++ frontend/src/pages/report/_components/index.ts
@@ -6,3 +6,4 @@ export * from './VideoSummary'
 export * from './Tag'
 export * from './TitledSection'
 export * from './GuestModal'
+export * from './UpdateModal'
